Fix valgrind failure

+++ tool-openssl/enc_test.cc
@@ -310,9 +310,11 @@ class EncComparisonTest : public ::testing::Test {
   }
 
   void TearDown() override {
-    RemoveFile(in_path);
-    RemoveFile(out_path_tool);
-    RemoveFile(out_path_openssl);
+    if (tool_executable_path != nullptr && openssl_executable_path != nullptr) {
+      RemoveFile(in_path);
+      RemoveFile(out_path_tool);
+      RemoveFile(out_path_openssl);
+    }
   }
 
   char in_path[PATH_MAX];
